Removed unused code and moved util classes into each package feature
